jamterm: align scrolling behavior with rio (thanks akw)
Subject: [PATCH] samterm: Simplify scrolling mechanics

This keeps Torigin messages contained to mostly center().  So now
center() takes an origin and line offset, to then move either locally
on the terminal (n > 0), or via messages to & from the host (n <= 0).

With this, Kup/Kdown go by 1/3 of the screen (as in rio), and
Kpgup/Kpgdown go by # of lines on the screen, so a Kpgup undoes a Kpgdown.

This also makes scrolling look smoother by flushing the display only
when necessary.