jamterm: scroll by 1 line with shift, extend selection with shift click (thanks akw)
Subject: [PATCH] samterm: Roll our own kbd proc

Now we can process Kshift characters so we may do Kshift+(Kup|Kdown)
for moving one line at a time (as in rio).  Also included is the
ability to "extend" a selection in the direction of the initial
selection, i.e. if we have:

	And, now, I don't mind a bit of a breeze, if anything I prefer
	it, but thon was aggressive.

If we were to select "mind" to "breeze", then a shift+left-click to
the right of "breeze" will extend the selection forward to that point,
and if we were to shift+left-click to a point before "mind", the
selection would reverse.